Simplify subclassing by adding ScriptBuilder.defineEventSource()
defineEventSource() sets up a global property and additionally creates
the `on()` event hook in its prototype. This allows us to get rid of the
CallbackHolder interface.

The new on() hook accepts a map of event bindings in addition to a single
event type and callback.

Also renamed Callbacks to EventMap.